+++ tests/testthat/test-tar-terra.R
@@ -1,7 +1,6 @@
 # test_that() #Included to make RStudio recognize this file as a test
 library(targets)
 targets::tar_test("tar_terra_rast() works", {
-    # geotargets::geotargets_option_set(gdal_raster_creation_options = c("COMPRESS=DEFLATE", "TFW=YES"))
     targets::tar_script({
         list(
             geotargets::tar_terra_rast(